some printf changes and be more verbose when a filelist entry is not correct
git-svn-id: file:///svn/rdup/trunk@1030 0f4cb41f-7f08-0410-98f9-8ec83d93c06f

+++ gfunc.c
@@ -197,7 +197,7 @@ entry_print_data(FILE *out, char n, struct entry *e)
 				fputc('0', out);
 				break;
 			}
-			fprintf(out, "%zd", (size_t)e->f_size);
+			fprintf(out, "%zd", e->f_size);
 			break;
 		case 'H': /* sha1 hash */
 			if (S_ISREG(e->f_mode))
@@ -327,7 +327,7 @@ gfunc_write(gpointer data, __attribute__((unused)) gpointer value, gpointer fp)
 
 	/* mode_path */
 	/* this is used to create our filelist */
-	fprintf((FILE*) fp, "%d %d %s", (int) e->f_mode, (int)e->f_name_size, 
+	fprintf((FILE*) fp, "%d %d %s", (int)e->f_mode, (int)e->f_name_size, 
 			e->f_name);
 	if (opt_null) {
 		fputc('\0', (FILE*)fp);
